Add keyboard shortcuts and cleanup
There are a lot of changes :eyes

In order to get keyboard shortcuts for all signals
we need them to be mounted, but a couple of them only became avaialble
once dropdown was expanded.

They've now been moved into the main tray and the design tweaked
so they fit in nicely.

The cleanup removes unnecessary divs, components, css and store state entry.